rename helpers -> gitHubTests
